Added skeleton placeholder and empty data view
