+++ website/components/YouMayLike/index.tsx
@@ -13,16 +13,12 @@ import supabase from "@/utils/supabase";
 
 import "swiper/css";
 import "swiper/css/effect-fade";
-import { InView } from "react-intersection-observer";
 
 const YouMayLike = () => {
-  const [isLoading, setIsLoading] = useState(true);
-  const [error, setError] = useState<any | null>(null);
   const [restaurants, setRestaurantsData] = useState<any | null>([]);
 
   useEffect(() => {
     const fetchData = async () => {
-      setIsLoading(true);
       try {
         const { data, error } = await supabase
           .from("restaurants")
@@ -30,14 +26,11 @@ const YouMayLike = () => {
           .eq("is_public", true)
           .limit(4);
 
-        if (error) throw error;
+        if (error) return error;
 
         setRestaurantsData(data);
       } catch (error) {
         console.error("Error while fetching restaurants data: ", error);
-        setError(error);
-      } finally {
-        setIsLoading(false);
       }
     };
 
@@ -54,64 +47,56 @@ const YouMayLike = () => {
             customClass="mb-28 mt-20"
           />
 
-          {isLoading ? (
-            <section>
-              <div className="py-20 text-center text-black/40 dark:text-white/70">
-                Loading...
-              </div>
-            </section>
-          ) : (
-            <div className="grid grid-cols-1 gap-4 xs:gap-10 lg:grid-cols-2">
-              {restaurants.map((item: any) => (
-                <div key={"may-like-" + item.id}>
-                  <InView triggerOnce>
-                    {({ inView, ref, entry }) => (
-                      <Link
-                        ref={ref}
-                        href={`/restaurants/${item.id}/menu`}
-                        className={`relative h-full cursor-pointer ${
-                          inView ? "animated-fade-y" : ""
-                        }`}
-                      >
-                        <Swiper
-                          modules={[Autoplay, EffectFade]}
-                          slidesPerView={1}
-                          loop={true}
-                          autoplay={true}
-                          effect="fade"
-                          className="sm:h-[25rem]"
-                        >
-                          {item.images.map((data: any) => (
-                            <div key={"image-" + data}>
-                              <SwiperSlide>
-                                <Image
-                                  src={data}
-                                  height={100}
-                                  width={100}
-                                  className="h-full w-full object-cover"
-                                  alt="item-image"
-                                />
-                              </SwiperSlide>
-                            </div>
-                          ))}
-                        </Swiper>
-                        <div className="absolute top-0 z-[1] h-full w-full bg-gradient-to-t from-black/70 to-transparent md:via-transparent">
-                          <div className="absolute bottom-2 z-[1] w-full px-3 text-white sm:bottom-5 sm:px-8">
-                            <p className="mb-2 border-b border-white border-opacity-30 pb-2 text-xl font-extrabold sm:text-2xl">
-                              {item.name}
-                            </p>
-                            <p className="text-sm text-gray-200 sm:text-base">
-                              {item.address}
-                            </p>
-                          </div>
-                        </div>
-                      </Link>
-                    )}
-                  </InView>
+          <div className="animated-fade-y grid grid-cols-1 gap-4 xs:gap-10 lg:grid-cols-2">
+            {restaurants.map((item: any) => (
+              <Link
+                key={"may-like-" + item.id}
+                href={
+                  "/restaurants/" +
+                  encodeURIComponent(
+                    item.name.toLowerCase().replace(/\s+/g, "-")
+                  ) +
+                  "-" +
+                  btoa(item.id.toString())
+                }
+                className="relative h-full cursor-pointer"
+              >
+                <Swiper
+                  modules={[Autoplay, EffectFade]}
+                  slidesPerView={1}
+                  loop={true}
+                  autoplay={true}
+                  effect="fade"
+                  className="sm:h-[25rem]"
+                >
+                  {item.images.map((data: any) => (
+                    <div key={"image-" + data}>
+                      <SwiperSlide>
+                        <Image
+                          src={data}
+                          height={100}
+                          width={100}
+                          className="h-full w-full object-cover"
+                          alt="item-image"
+                          loading="lazy"
+                        />
+                      </SwiperSlide>
+                    </div>
+                  ))}
+                </Swiper>
+                <div className="absolute top-0 z-[1] h-full w-full bg-gradient-to-t from-black/70 to-transparent md:via-transparent">
+                  <div className="absolute bottom-2 z-[1] w-full px-3 text-white sm:bottom-5 sm:px-8">
+                    <p className="mb-2 border-b border-white border-opacity-30 pb-2 text-xl font-extrabold sm:text-2xl">
+                      {item.name}
+                    </p>
+                    <p className="text-sm text-gray-200 sm:text-base">
+                      {item.address}
+                    </p>
+                  </div>
                 </div>
-              ))}
-            </div>
-          )}
+              </Link>
+            ))}
+          </div>
         </div>
       </section>
     </>
