+++ website/types/card-item.ts
@@ -7,7 +7,6 @@ export type ItemProps = {
   video?: string;
   price: number;
   rating: number;
-  // `restaurants` is an object not an array, cause it's a relational object of menus data
   menus: {
     id: number,
     restaurants: {
